+++ Coveralls.Lib/Coveralls.Lib.csproj
@@ -64,7 +64,8 @@
     <Compile Include="ParserTypes.cs" />
     <Compile Include="Parser\BaseParser.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="StringExtensions.cs" />
+    <Compile Include="PathTools.cs" />
+    <Compile Include="StringTools.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
